Renaming from Silex to Pimple
