+++ src/Container/ClientsContainer.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Predis\Silex\Container;
+namespace Predis\Pimple\Container;
 
 use BadMethodCallException;
 use Pimple\Container;
